requirements.txt
Django==1.8.1
djangorestframework==3.1.1
gunicorn==19.3.0
psycopg2==2.6
django-postgrespool==0.3.0
dj-database-url==0.3.0
selenium==2.46.0
flake8==2.4.1
factory-boy==2.4.1
Paver==1.2.4
